+++ src/main/java/vo/AccountBillItemVO.java
@@ -15,11 +15,11 @@ public class AccountBillItemVO {
 	/** 银行账户 */
 	public String accountID;
 	/** 转账金额 */
-	public int money;
+	public double money;
 	/** 备注 */
 	public String remark;
 
-	public AccountBillItemVO(String accountID, int money, String remark) {
+	public AccountBillItemVO(String accountID, double money, String remark) {
 		this.accountID = accountID;
 		this.money = money;
 		this.remark = remark;
